+++ packages/ui-scaffold/src/components/restrictedAreas/index.ts
@@ -1,3 +1,0 @@
-export { RestrictedAreas } from "./restrictedAreas.ui";
-export { RestrictedAreasWidget } from "./restrictedAreas.widget";
-export { useRestrictedAreasScript } from "./restrictedAreas.script";