+++ packages/ui-scaffold/src/components/restrictedInfo/index.ts
@@ -1,0 +1,3 @@
+export { RestrictedInfo } from "./restrictedInfo.ui";
+export { RestrictedInfoWidget } from "./restrictedInfo.widget";
+export { useRestrictedInfoScript } from "./restrictedInfo.script";