+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @orderly.network/cli
 
+## 0.3.22
+
+### Patch Changes
+
+- 74c2c433e: fix bugs
+- Updated dependencies [74c2c433e]
+  - @orderly.network/core@2.0.6
+
+## 0.3.22-alpha.0
+
+### Patch Changes
+
+- fix bugs
+- Updated dependencies
+  - @orderly.network/core@2.0.6-alpha.0
+
 ## 0.3.21
 
 ### Patch Changes