+++ packages/default-solana-adapter/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @orderly.network/default-solana-adapter
 
+## 2.0.6
+
+### Patch Changes
+
+- 74c2c433e: fix bugs
+- Updated dependencies [74c2c433e]
+  - @orderly.network/core@2.0.6
+  - @orderly.network/types@2.0.6
+
+## 2.0.6-alpha.0
+
+### Patch Changes
+
+- fix bugs
+- Updated dependencies
+  - @orderly.network/core@2.0.6-alpha.0
+  - @orderly.network/types@2.0.6-alpha.0
+
 ## 2.0.5
 
 ### Patch Changes