+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @orderly.network/hooks
 
+## 2.0.6
+
+### Patch Changes
+
+- 74c2c433e: fix bugs
+- Updated dependencies [74c2c433e]
+  - @orderly.network/default-solana-adapter@2.0.6
+  - @orderly.network/core@2.0.6
+  - @orderly.network/default-evm-adapter@2.0.6
+  - @orderly.network/net@2.0.6
+  - @orderly.network/perp@4.0.5
+  - @orderly.network/types@2.0.6
+  - @orderly.network/utils@2.0.6
+  - @orderly.network/web3-provider-ethers@2.0.6
+
+## 2.0.6-alpha.0
+
+### Patch Changes
+
+- fix bugs
+- Updated dependencies
+  - @orderly.network/default-solana-adapter@2.0.6-alpha.0
+  - @orderly.network/core@2.0.6-alpha.0
+  - @orderly.network/default-evm-adapter@2.0.6-alpha.0
+  - @orderly.network/net@2.0.6-alpha.0
+  - @orderly.network/perp@4.0.5-alpha.0
+  - @orderly.network/types@2.0.6-alpha.0
+  - @orderly.network/utils@2.0.6-alpha.0
+  - @orderly.network/web3-provider-ethers@2.0.6-alpha.0
+
 ## 2.0.5
 
 ### Patch Changes